* Converted the onboard code (both MATLAB and C) to use the new GPS datatypes that were updated in the simulator within the last few commits.  * Added a new status bit for the boat for whether the RC transmitter is calibrated.  * RC_CHANNELS_SCALED is now only transmit once the transmitter is calibrated.  * RC_CHANNELS_RAW now properly transmits the uptime of the incoming PWM so that units are now correct.  * gpsData.alt cannot be relied on as of now as it is not ever filled with valid data. It should be removed as it's defunct.  * Updated the documentation for the SYSTEM_STATUS message.

+++ Code/basic_model/clib/nmea2000.h
@@ -75,10 +75,10 @@ uint8_t ParsePgn128259(uint8_t data[8], uint8_t *seqId, float *waterSpeed);
 uint8_t ParsePgn128267(uint8_t data[8], uint8_t *seqId, float *waterDepth, float *offset);
 
 // Units are seqId: none, latitude: radians (+ north), longitude: radians (+ east).
-uint8_t ParsePgn129025(uint8_t data[8], float *latitude, float *longitude);
+uint8_t ParsePgn129025(uint8_t data[8], int32_t *latitude, int32_t *longitude);
 
 // Units are seqId: none, cogRef: 0 (True), 1 (magnetic), cog: radians eastward from north, sog: m/s
-uint8_t ParsePgn129026(uint8_t data[8], uint8_t *seqId, uint8_t *cogRef, float *cog, float *sog);
+uint8_t ParsePgn129026(uint8_t data[8], uint8_t *seqId, uint8_t *cogRef, uint16_t *cog, uint16_t *sog);
 
 // Units are seqId: none, airSpeed: m/s, direction: radians eastward from north.
 uint8_t ParsePgn130306(uint8_t data[8], uint8_t *seqId, float *airSpeed, float *direction);
